bring back "Relations" field when creating a server
Somehow this field disappeared from our API docs. I found out that it is just a optional field. If it is set, all arrays of its sub-fields have to be filled with either empty "[]" or some elements. This means if one sub-field is filled and others are empty, those empty fields still have to be presented (with "[]" as values).